+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 - CLI set_defaults option to allow overriding of defaults.
 - CLI return_parser option to ease inclusion in documentation.
 - save_path_content attribute to save paths content on config save.
+- New `link_arguments` method to derive an argument value from another.
 
 Changed
 ^^^^^^^